+++ scripts/install_ubuntu_deps.sh
@@ -10,11 +10,20 @@ sudo apt-get install -y \
   kmod \
   libgrpc++-dev \
   libprotobuf-dev \
+  libxkbcommon-dev \
   libyaml-cpp-dev \
   ninja-build \
   pkg-config \
   protobuf-compiler \
-  protobuf-compiler-grpc
+  protobuf-compiler-grpc \
+  qml6-module-qtqml-workerscript \
+  qml6-module-qtquick \
+  qml6-module-qtquick-controls \
+  qml6-module-qtquick-layouts \
+  qml6-module-qtquick-templates \
+  qml6-module-qtquick-window \
+  qt6-base-dev \
+  qt6-declarative-dev
 
 cmake --version
 g++ --version